@pbogdan 2017-06-29
FYI I believe this might be the PR @Ericson2314 was talking about https://github.com/NixOS/nixpkgs/pull/26554