@orivej 2017-06-30
If bash parser in `staging` needs to be replaced, I published a simpler C++ one at https://github.com/orivej/expand-compiler-args