@ryantrinkle 2017-06-30
@orivej I have no complaint if you'd rather use your code.  My goal was just to solve the issue while keeping the code, dependencies, and bootstrapping logic as simple as possible.  I'm not familiar enough with C++'s toolchain to know whether it would present any difficulties.  I did have to restrain myself from trying to do it in Haskell, though—attoparsec would've made this quite a bit nicer!

BTW, the file I've been testing with is [responsefile.txt](https://github.com/NixOS/nixpkgs/files/1116415/responsefile.txt).