@orivej 2017-06-30
I had not considered that C++ toolchain is not available as soon as C toolchain is. I appreciate the work you did to integrate compiled rsp parser into bootstrap process, thank you! Yet `parseResponseFile.c` and `expandResponseParams()` could be much simpler, so I suggest that you replace them with my implementation, which is now in C.