@0xABAB 2017-06-30
I preferred the C++ version, assuming it had the same semantics.