@orivej 2017-07-01
If C++ is an option for early Nix bootstrap, I made updated version available at the [cpp branch](https://github.com/orivej/expand-compiler-args/blob/cpp/expand-compiler-args.cpp). The only semantic change was about handling incorrect .rsp files as gcc, e.g. `good arg and 'unclosed with trailing\` is now parsed as `normal arg and unclosed arg with trailing`.

Current timings of parsing [responsefile.txt](https://github.com/NixOS/nixpkgs/files/1116415/responsefile.txt) with C, `staging`, and `master` versions of `expandResponseParams` are: 0m0.013s, 0m2.110s, and 3m25.268s. (Actual C parser completes under 2ms, the rest must be Bash overhead.)