@orivej 2017-07-01
Alternatively, bash version could be optimized somewhat further:
https://github.com/orivej/expand-compiler-args/commit/d77495a8b7d295b0f4f10d7e6d9c225f0e8640df reduces `responsefile.txt` parsing time from 2s to 1.2s, `ghc_7.rsp` from 1s to 0.6s.