@0xABAB 2017-07-01
Can we get anything which improves the situation *now* as opposed to what might be nice to have some later date in the future? The initial 75,000 factor of improvement seems to be interesting enough to include regardless of whether it could be made faster in another PR. 

At least that's the process I would like to see.