@orivej 2017-07-01
> Can we get anything which improves the situation now as opposed to what might be nice to have some later date in the future?
Yes, merge nixpkgs `staging` branch into `master`.
> The initial 75,000 factor of improvement seems to be interesting enough to include regardless of whether it could be made faster in another PR.
This PR is for the `staging` branch, as it should be because it causes everything to be rebuilt. Merging it will reduce parsing time of the provided sample from 2s to 0.02s.