@0xABAB 2017-07-01
OK, so we are now just waiting for someone to push the merge button then, unless there are more comments from others.