@orivej 2017-07-01
No, merging this PR has nothing to do with merging `staging` into `master`. This PR is about speeding up `staging`.