@0xABAB 2017-07-01
@orivej I did not say there was, since the top says `ryantrinkle  wants to merge 1 commit into NixOS:staging from obsidiansystems:response-file-parsing-speed`. I am well aware as to what the role is of `staging`.