@orivej 2017-07-01
I'd wait for @ryantrinkle on https://github.com/obsidiansystems/nixpkgs/pull/3, and for people here to say if they want this additional speedup or if they do not want this additional complexity. For my purposes current `staging` is good enough.