@Ericson2314 2017-07-02
I think C++ should be fine.  Also, if I recall correctly, there is some `langCC` attribute or similar on compilers that can be used to tell whether the compiler is supposed to support it.