@orivej 2017-07-04
This PR is ready for merge.