@nh2 2017-07-24
I found that even after this PR merged (nixpkgs commit aad2a2132c4 for me), `ld-wrapper` invocations can still take 10 seconds of CPU time in bash for each executable for large Haskell projects.