@nh2 2017-08-17
> invocations can still take 10 seconds of CPU time in bash for each executable for large Haskell projects

Fixed in #27657.